The mac libraries are now versioned and will download automatically.

+++ project/mac-xcode/fetchDepends.sh
@@ -1,7 +1,13 @@
 #!/bin/sh
 cd ../../lib
-rm -rf mac
-mkdir mac
-cd mac
-curl -z maclib.zip http://boxedwine.org/v/depend/mac/2/maclib.zip --output maclib.zip
-unzip maclib.zip
+libVersion=$(< mac/version.txt)
+if [ $libVersion -eq 3 ]; then
+  echo "Mac Lib up to date."
+else 
+  echo "Update Mac Lib"
+  rm -rf mac
+  mkdir mac
+  cd mac
+  curl -z maclib.zip http://boxedwine.org/v/depend/mac/3/maclib.zip --output maclib.zip
+  unzip maclib.zip
+fi
